refactor: move fuel upload log model
